pyproject.toml
[tool.poetry]
name = "dmarc-report-aggregator"
version = "1.0.0"
description = "Receives DMARC reports via SMTP(S) and visualizes them via HTTP."
authors = ["Maximilian Haye <maxh@freitagsrunde.org>"]
packages = [{ include = "dmarc_report_aggregator" }]

[tool.poetry.dependencies]
python = "^3.11"
aiosmtpd = "^1.4.4.post2"
dkimpy = { version = "^1.1.4", extras = ["asyncio"] }
aiohttp = "^3.8.5"
aiodns = "^3.0.0"
aiosqlite = "^0.19.0"
aiohttp-jinja2 = "^1.5.1"
pydantic = "^2.1.1"
ldap3 = "^2.9.1"
pydantic-settings = "^2.0.2"
aiohttp-session = { extras = ["secure"], version = "^2.12.0" }
defusedxml = "^0.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
aiomonitor = "^0.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
